requirements.txt
apache-airflow==2.7.1
googlemaps==4.10.0
pandas==2.0.3
scikit-learn==1.3.0
psycopg2-binary==2.9.7
minio==7.1.15
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2 